updating dependencies to not be global for sregistry client
