+++ requirements.txt
@@ -1,6 +1,4 @@
 python-dateutil
-demjson
 requests
 requests-toolbelt
 retrying
-simplejson
