+++ sregistry/auth/utils.py
@@ -1,8 +1,8 @@
 '''
 
-Copyright (C) 2017 The Board of Trustees of the Leland Stanford Junior
+Copyright (C) 2017-2018 The Board of Trustees of the Leland Stanford Junior
 University.
-Copyright (C) 2016-2017 Vanessa Sochat.
+Copyright (C) 2016-2018 Vanessa Sochat.
 
 This program is free software: you can redistribute it and/or modify it
 under the terms of the GNU Affero General Public License as published by
@@ -20,19 +20,8 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 '''
 
 from sregistry.logger import bot
-from sregistry.utils import (
-    read_json,
-    write_json
-)
-
-from datetime import datetime, timezone
 import base64
-import hashlib
-import hmac
-import json
 import os
-import pwd
-import requests
 import sys
 
 
@@ -42,32 +31,6 @@ def _encode(item):
     if not isinstance(item,bytes):
         item = item.encode('utf-8')
     return item
-
-
-def generate_signature(payload, secret):
-    '''use an endpoint specific payload and client secret to generate
-    a signature for the request'''
-    payload = _encode(payload)
-    secret = _encode(secret)
-    return hmac.new(secret, digestmod=hashlib.sha256,
-                    msg=payload).hexdigest()
-
-
-def generate_timestamp():
-    ts = datetime.now(timezone.utc)
-    return ts.strftime('%Y%m%dT%HZ')
-
-
-def generate_credential(s):
-    '''basic_auth_header will return a base64 encoded header object to
-    :param username: the username
-    '''
-    if sys.version_info[0] >= 3:
-        s = bytes(s, 'utf-8')
-        credentials = base64.b64encode(s).decode('utf-8')
-    else:
-        credentials = base64.b64encode(s)
-    return credentials
 
 
 def basic_auth_header(username, password):
@@ -89,16 +52,3 @@ def basic_auth_header(username, password):
         credentials = base64.b64encode(s)
     auth = {"Authorization": "Basic %s" % credentials}
     return auth
-
-
-def generate_header_signature(secret, payload, request_type):
-    '''Authorize a client based on encrypting the payload with the client
-       secret, timestamp, and other metadata
-     '''
-
-    # Use the payload to generate a digest   push|collection|name|tag|user
-    timestamp = generate_timestamp()
-    credential = "%s/%s" %(request_type,timestamp)
-
-    signature = generate_signature(payload,secret)
-    return "SREGISTRY-HMAC-SHA256 Credential=%s,Signature=%s" %(credential,signature)
