+++ sregistry/main/registry/delete.py
@@ -27,7 +27,7 @@ from sregistry.utils import ( parse_image_name, remove_uri )
 import sys
 import os
 
-from sregistry.auth import (
+from sregistry.main.registry.auth import (
     generate_signature,
     generate_credential,
     generate_timestamp
