+++ sregistry/version.py
@@ -19,7 +19,7 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 '''
 
-__version__ = "0.0.61"
+__version__ = "0.0.62"
 AUTHOR = 'Vanessa Sochat'
 AUTHOR_EMAIL = 'vsochat@stanford.edu'
 NAME = 'sregistry'
@@ -29,38 +29,39 @@ DESCRIPTION = "Command line tool for working with singularity-hub and registry."
 LICENSE = "LICENSE"
 
 INSTALL_REQUIRES = (
-
-    ('dateutils', {'min_version': "0.6.6"}),
-    ('demjson', {'exact_version': "2.2.4"}),
-    ('python-dateutil', {'exact_verison': "2.5.3"}),
     ('requests', {'exact_version': '2.18.4'}),
-    ('requests-toolbelt', {'exact_version': '0.8.0'}),
     ('retrying', {'exact_version': '1.3.3'}),
     ('pygments', {'min_version': '2.1.3'}),
-    ('sqlalchemy', {'min_version': None}),
-    ('oauth2client', {'exact_version': '3.0'})
+    ('sqlalchemy', {'min_version': None})
 )
 
 # Submodule Requirements
 
 INSTALL_REQUIRES_GLOBUS = (
-
     ('globus-sdk[jwt]', {'exact_version': '1.3.0'}),
+)
+
+INSTALL_REQUIRES_REGISTRY = (
+    ('requests-toolbelt', {'exact_version': '0.8.0'}),
+    ('dateutils', {'min_version': "0.6.6"}),
+    ('python-dateutil', {'min_verison': "2.5.3"})
 )
 
 INSTALL_REQUIRES_DROPBOX = (
 )
 
 INSTALL_REQUIRES_GOOGLE_STORAGE = (
-    ('google-cloud-storage', {'min_version': '1.4.0'}),
-
+    ('oauth2client', {'min_version': '3.0'}),
+    ('google-cloud-storage', {'min_version': '1.4.0'})
 )
 
 INSTALL_REQUIRES_GOOGLE_DRIVE = (
+    ('oauth2client', {'min_version': '3.0'}),
     ('google-api-python-client', {'min_version': '1.6.4'}),
 )
 
 INSTALL_REQUIRES_ALL = (INSTALL_REQUIRES +
                         INSTALL_REQUIRES_DROPBOX + 
+                        INSTALL_REQUIRES_REGISTRY +
                         INSTALL_REQUIRES_GOOGLE_STORAGE +
                         INSTALL_REQUIRES_GOOGLE_DRIVE)
